+++ acme/resources/ACP.py
@@ -13,6 +13,7 @@ from ..helpers.TextTools import simpleMatch
 from ..helpers.TextTools import findXPath
 from ..etc.Types import AttributePolicyDict, ResourceTypes, Result, Permission, JSON
 from ..etc.ResponseStatusCodes import BAD_REQUEST
+from ..etc.Constants import Constants
 from ..services import CSE
 from ..services.Logging import Logging as L
 from ..resources.Resource import Resource
@@ -21,6 +22,8 @@ from ..resources.AnnounceableResource import AnnounceableResource
 
 class ACP(AnnounceableResource):
 	""" AccessControlPolicy (ACP) resource type """
+
+	_riTyMapping = Constants.attrRiTyMapping
 
 	_allowedChildResourceTypes:list[ResourceTypes] = [ ResourceTypes.SUB ] # TODO Transaction to be added
 	""" The allowed child-resource types. """
@@ -56,6 +59,8 @@ class ACP(AnnounceableResource):
 					   create:Optional[bool] = False) -> None:
 		super().__init__(ResourceTypes.ACP, dct, pi, create = create, inheritACP = True, rn = rn)
 
+		self._addToInternalAttributes(self._riTyMapping)
+
 		self.setAttribute('pv/acr', [], overwrite = False)
 		self.setAttribute('pvs/acr', [], overwrite = False)
 
@@ -84,6 +89,27 @@ class ACP(AnnounceableResource):
 
 		_checkAcod(findXPath(dct, f'{ResourceTypes.ACPAnnc.tpe()}/pv/acr'))
 		_checkAcod(findXPath(dct, f'{ResourceTypes.ACPAnnc.tpe()}/pvs/acr'))
+
+		# Get types for the acor members. Ignore if not found
+		# This is an optimization used later in case there is a group in acor
+		riTyDict = {}
+
+		def _getAcorTypes(pv:JSON) -> None:
+			if pv:
+				for acr in pv.get('acr', []):
+					if (acor := acr.get('acor')):
+						for o in acor:
+							try:
+								r = CSE.dispatcher.retrieveResource(o)
+								riTyDict[o] = r.ty		
+							except:
+								# ignore any errors here. The acor might not be a resource yet
+								continue
+
+		_getAcorTypes(self.getFinalResourceAttribute('pv', dct))
+		_getAcorTypes(self.getFinalResourceAttribute('pvs', dct))
+		self.setAttribute(ACP._riTyMapping, riTyDict)
+
 
 
 	def deactivate(self, originator:str) -> None:
@@ -183,10 +209,14 @@ class ACP(AnnounceableResource):
 				# TODO support acod/specialization
 
 			# Check originator
-			if 'all' in acr['acor'] or originator in acr['acor'] or requestedPermission == Permission.NOTIFY:
+			if self._checkAcor(acr['acor'], originator):
 				return True
-			if any([ simpleMatch(originator, a) for a in acr['acor'] ]):	# check whether there is a wildcard match
-				return True
+
+
+			# if 'all' in acr['acor'] or originator in acr['acor'] or requestedPermission == Permission.NOTIFY:
+			# 	return True
+			# if any([ simpleMatch(originator, a) for a in acr['acor'] ]):	# check whether there is a wildcard match
+			# 	return True
 		return False
 
 
@@ -204,10 +234,42 @@ class ACP(AnnounceableResource):
 		for p in self['pvs/acr']:
 			if requestedPermission & p['acop'] == 0:	# permission not fitting at all
 				continue
-			# TODO check acod in pvs
-			if 'all' in p['acor'] or originator in p['acor']:
+
+			# Check originator
+			if self._checkAcor(p['acor'], originator):
 				return True
-			if any([ simpleMatch(originator, a) for a in p['acor'] ]):	# check whether there is a wildcard match
-				return True
+
+			# if 'all' in p['acor'] or originator in p['acor']:
+			# 	return True
+			# if any([ simpleMatch(originator, a) for a in p['acor'] ]):	# check whether there is a wildcard match
+			# 	return True
 		return False
 
+
+	def _checkAcor(self, acor:list[str], originator:str) -> bool:
+
+		# Check originator
+		if 'all' in acor or \
+			originator in acor:
+			# or requestedPermission == Permission.NOTIFY:	# TODO not sure whether this is correct
+			return True
+		
+		# Iterrate over all acor entries for either a group check or a wildcard check
+		_riTypes = self.attribute(ACP._riTyMapping)
+		for a in acor:
+
+			# Check for group. If the originator is a member of the group, then the originator has access
+			if _riTypes.get(a) == ResourceTypes.GRP:
+				try:
+					if originator in CSE.dispatcher.retrieveResource(a).mid:
+						L.isDebug and L.logDebug(f'Originator found in group member')
+						return True
+				except Exception as e:
+					L.logErr(f'GRP resource not found for ACP check: {a}')
+					continue # Not much that we can do here
+
+			# Otherwise Check for wildcard match
+			if simpleMatch(originator, a):
+				return True
+		
+		return False